Check for qmldir_compile and use it when compiling
When compiling qml, when looking for imports etc. the qml files of plugins are
needed. For the test they are in the .so plugin so they aren't listed in
qmldir_compile. When running/loading the .qmc files are used and listed in
qmldir.

Would be nice to do the other way around(qmldir_loading) so apps not using
compiled qml files keep on working but this would need a change in qtdeclarative
or copying code from there.